+++ docs/ReleaseNotes.md
@@ -1,6 +1,10 @@
 Release Notes for Version 13.0
 =============================
 
+* Added the ability to pass in multiple indexes as the first parameter to IString.formatChoice(). This allows you to format 
+  choice strings where there are two or more plurals in the string at the same time. For example, if your string is "There
+  are N items on P pages", you can get the locale-sensitive plurals of "items" and "pages" correct with one call.
+  
 Build 001
 -------